Fix merge error in make_app_test
